requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.43
psycopg2-binary==2.9.10
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
email-validator==2.2.0
pydantic==2.9.2
pydantic[email]
alembic==1.16.5
python-dotenv==1.1.1